Feat/bookmarks

+++ src/repo/repo.service.ts
@@ -17,7 +17,7 @@ export class RepoService {
 
   async createRepo(
     userId: string,
-    { name, description, isPublic }: CreateRepoDto,
+    { name, description, tags, isPublic }: CreateRepoDto,
   ) {
     const user = await this.prisma.user.findUnique({ where: { id: userId } });
 
@@ -26,7 +26,13 @@ export class RepoService {
     }
 
     const repo = await this.prisma.repo.create({
-      data: { name, description, isPublic, user: { connect: { id: userId } } },
+      data: {
+        name,
+        description,
+        isPublic,
+        tags,
+        user: { connect: { id: userId } },
+      },
       include: { user: false },
     });
 
@@ -59,6 +65,80 @@ export class RepoService {
     return repo;
   }
 
+  async bookmarkRepo(userId: string, repoId: string) {
+    const repo = await this.prisma.repo.findUnique({
+      where: { id: repoId },
+    });
+
+    if (!repo) {
+      throw new NotFoundException('Repo not found');
+    }
+
+    const bookmarked = await this.prisma.bookmark.create({
+      data: { repoId, user: { connect: { id: userId } } },
+    });
+
+    return bookmarked;
+  }
+
+  async unbookmarkRepo(userId: string, repoId: string) {
+    const repo = await this.prisma.repo.findUnique({
+      where: { id: repoId },
+    });
+
+    if (!repo) {
+      throw new NotFoundException('Repo not found');
+    }
+
+    await this.prisma.$transaction([
+      this.prisma.user.update({
+        where: { id: userId },
+        data: { bookmarks: { disconnect: { id: repoId } } },
+      }),
+      this.prisma.bookmark.delete({ where: { id: repoId } }),
+    ]);
+
+    return;
+  }
+
+  async getBookmarks(userId: string) {
+    const bookmarks = await this.prisma.bookmark.findMany({
+      where: { userId },
+      select: { repoId: true },
+    });
+
+    if (!bookmarks) {
+      throw new NotFoundException('No bookmarks found');
+    }
+
+    const bookmarkIds: string[] = [];
+
+    bookmarks.forEach((bookmark) => bookmarkIds.push(bookmark.repoId));
+
+    const bookmarkedRepos = await this.prisma.repo.findMany({
+      where: { id: { in: bookmarkIds } },
+    });
+
+    return bookmarkedRepos;
+  }
+
+  async getBookmarksRepoIds(userId: string) {
+    const bookmarks = await this.prisma.bookmark.findMany({
+      where: { userId },
+      select: { repoId: true },
+    });
+
+    if (!bookmarks) {
+      throw new NotFoundException('No bookmarks found');
+    }
+
+    const bookmarkIds: string[] = [];
+
+    bookmarks.forEach((bookmark) => bookmarkIds.push(bookmark.repoId));
+
+    return { repoIds: bookmarkIds };
+  }
+
   async deleteUserRepo(userId: string, repoId: string) {
     const repo = await this.prisma.repo.findUnique({
       where: { id: repoId, userId },
@@ -71,23 +151,35 @@ export class RepoService {
       );
     }
 
-    //For storing file names
-    const fileNames: string[] = [];
+    //If user has files
+    if (repo.files.length > 0) {
+      //For storing file names
+      const fileNames: string[] = [];
 
-    repo.files.forEach((file) => fileNames.push(file.publicName));
+      repo.files.forEach((file) => fileNames.push(file.publicName));
 
-    //Delete all files from cloudinary, to be implemented
-    await this.cloudinary.deleteFiles(fileNames);
+      //Delete all file relations to repo and delete repo
+      await this.prisma.$transaction([
+        this.prisma.repo.update({
+          where: { id: repoId },
+          data: { files: { deleteMany: {} } },
+          include: { files: true },
+        }),
+        this.prisma.user.update({
+          where: { id: userId },
+          data: { Repo: { delete: { id: repoId } } },
+        }),
+      ]);
 
-    //Delete all file relations to repo and delete repo
-    await this.prisma.$transaction([
-      this.prisma.repo.update({
+      //Delete all files from cloudinary, to be implemented
+      await this.cloudinary.deleteFiles(fileNames);
+    } else {
+      //Delete only the repo
+      await this.prisma.user.update({
         where: { id: userId },
-        data: { files: { deleteMany: {} } },
-        include: { files: true },
-      }),
-      this.prisma.repo.delete({ where: { id: repoId } }),
-    ]);
+        data: { Repo: { delete: { id: repoId } } },
+      });
+    }
 
     return;
   }
